+++ extras/beast/test/test_allocator.hpp
@@ -17,42 +17,47 @@ namespace test {
 
 struct test_allocator_info
 {
+    std::size_t id;
     std::size_t ncopy = 0;
     std::size_t nmove = 0;
+    std::size_t nmassign = 0;
+    std::size_t ncpassign = 0;
     std::size_t nselect = 0;
+
+    test_allocator_info()
+        : id([]
+            {
+                static std::atomic<std::size_t> sid(0);
+                return ++sid;
+            }())
+    {
+    }
 };
 
-template<class T,
-    bool Equal, bool Assign, bool Move, bool Swap, bool Select>
+template<class T, bool Equal, bool Assign, bool Move, bool Swap, bool Select>
 class test_allocator;
 
-template<class T,
-    bool Equal, bool Assign, bool Move, bool Swap, bool Select>
+template<class T, bool Equal, bool Assign, bool Move, bool Swap, bool Select>
 struct test_allocator_base
 {
 };
 
-template<class T,
-    bool Equal, bool Assign, bool Move, bool Swap>
-struct test_allocator_base<
-    T, Equal, Assign, Move, Swap, true>
+template<class T, bool Equal, bool Assign, bool Move, bool Swap>
+struct test_allocator_base<T, Equal, Assign, Move, Swap, true>
 {
     static
     test_allocator<T, Equal, Assign, Move, Swap, true>
     select_on_container_copy_construction(test_allocator<
         T, Equal, Assign, Move, Swap, true> const& a)
     {
-        return test_allocator<T,
-            Equal, Assign, Move, Swap, true>{};
+        return test_allocator<T, Equal, Assign, Move, Swap, true>{};
     }
 };
 
-template<class T,
-    bool Equal, bool Assign, bool Move, bool Swap, bool Select>
+template<class T, bool Equal, bool Assign, bool Move, bool Swap, bool Select>
 class test_allocator : public test_allocator_base<
         T, Equal, Assign, Move, Swap, Select>
 {
-    std::size_t id_;
     std::shared_ptr<test_allocator_info> info_;
 
     template<class, bool, bool, bool, bool, bool>
@@ -73,24 +78,16 @@ public:
     template<class U>
     struct rebind
     {
-        using other = test_allocator<
-            U, Equal, Assign, Move, Swap, Select>;
+        using other = test_allocator<U, Equal, Assign, Move, Swap, Select>;
     };
 
     test_allocator()
-        : id_([]
-            {
-                static std::atomic<
-                    std::size_t> sid(0);
-                return ++sid;
-            }())
-        , info_(std::make_shared<test_allocator_info>())
+        : info_(std::make_shared<test_allocator_info>())
     {
     }
 
     test_allocator(test_allocator const& u) noexcept
-        : id_(u.id_)
-        , info_(u.info_)
+        : info_(u.info_)
     {
         ++info_->ncopy;
     }
@@ -98,17 +95,31 @@ public:
     template<class U>
     test_allocator(test_allocator<U,
             Equal, Assign, Move, Swap, Select> const& u) noexcept
-        : id_(u.id_)
-        , info_(u.info_)
+        : info_(u.info_)
     {
         ++info_->ncopy;
     }
 
     test_allocator(test_allocator&& t)
-        : id_(t.id_)
-        , info_(t.info_)
+        : info_(t.info_)
     {
         ++info_->nmove;
+    }
+
+    test_allocator&
+    operator=(test_allocator const& u) noexcept
+    {
+        info_ = u.info_;
+        ++info_->ncpassign;
+        return *this;
+    }
+
+    test_allocator&
+    operator=(test_allocator&& u) noexcept
+    {
+        info_ = u.info_;
+        ++info_->nmassign;
+        return *this;
     }
 
     value_type*
@@ -127,7 +138,7 @@ public:
     bool
     operator==(test_allocator const& other) const
     {
-        return id_ == other.id_ || Equal;
+        return id() == other.id() || Equal;
     }
 
     bool
@@ -139,7 +150,7 @@ public:
     std::size_t
     id() const
     {
-        return id_;
+        return info_->id;
     }
 
     test_allocator_info const*